extract ChunkDBManager
